Added server sided automatic demo recording
